worker/requirements.txt
confluent-kafka==2.3.0
pillow==10.0.0
opencv-python==4.8.1.78
numpy==1.24.3
